split .h and .cpp
