+++ deep8/nodes/Abs.h
@@ -1,24 +1,14 @@
 #ifndef DEEP8_ABS_H
 #define DEEP8_ABS_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
 /**
  * y = |x|
  */
 
-template <typename T>
-struct AbsBackwardExpr {
-    inline T operator()(T outputGrad, T input) const {
-        if (input > T(0)) {
-            return outputGrad;
-        } else if (input < T(0)) {
-            return -outputGrad;
-        } else {
-            return 0;
-        }
-    }
-};
 
 #ifdef HAVE_CUDA
 
@@ -55,34 +45,15 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the Abs Function needs only 1 input");
-
-        this->outputShape = this->inputs[0]->outputShape;
-    }
+    void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-        auto eigenDevice = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-        eTVec(output).device(*eigenDevice) = eTVec(inputs[0]).abs();
-    }
-
+    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
     void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
                      const Tensor<T> *output,
                      const Tensor<T> *outputGradient,
                      size_t index,
-                     Tensor<T> *iGradient) override {
-        if (0 != index) {
-            DEEP8_RUNTIME_ERROR("the index of Abs backwardCPU is error");
-        }
-
-        auto eigenDevice = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-        eTVec(iGradient).device(*eigenDevice) += eTVec(outputGradient).binaryExpr(eTVec(inputs[0]), AbsBackwardExpr<T>());
-    }
+                     Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
