+++ deep8/nodes/DivideScalar.h
@@ -1,6 +1,9 @@
 #ifndef DEEP8_DIVIDESCALAR_H
 #define DEEP8_DIVIDESCALAR_H
 
+#include "Function.h"
+
+
 namespace Deep8 {
 
 /*****************************************************************************/
@@ -40,51 +43,18 @@ public:
 		check();
 	}
 
-	void check() override {
-		checkImpl<T>();
-	}
+	void check() override;
 
 protected:
 	template <typename real>
-	void checkImpl() {
-		Function<real>::check();
+	void checkImpl();
 
-		DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the inputs size must be 1 in DivideScalar Function");
-		DEEP8_ARGUMENT_CHECK(0 != scalar, "the divide scalar can no be 0");
-
-		this->outputShape = this->inputs[0]->outputShape;
-	}
-
-#ifdef HAVE_HALF
-	template <>
-	void checkImpl<half>() {
-		Function<half>::check();
-
-		DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the inputs size must be 1 in DivideScalar Function");
-		DEEP8_ARGUMENT_CHECK(0 != __half2float(scalar), "the divide scalar can no be 0");
-
-		this->outputShape = this->inputs[0]->outputShape;
-	}
-#endif // HAVE_HALF
-
-
-	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-		auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-		eTVec(output).device(*device) = eTVec(inputs[0]) / scalar;
-	}
-	
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
 	void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
 		const Tensor<T> *output,
 		const Tensor<T> *outputGradient,
 		size_t index,
-		Tensor<T> *iGradient) override {
-		DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
-
-		auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
-
-		eTVec(iGradient).device(*device) += eTVec(outputGradient) / scalar;
-	}
+		Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
