+++ deep8/nodes/Exp.h
@@ -1,6 +1,9 @@
 #ifndef DEEP8_EXP_H
 #define DEEP8_EXP_H
 
+#include "Function.h"
+
+
 namespace Deep8 {
 
 /**
@@ -38,33 +41,16 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the Exp Function needs only 1 input");
-
-        this->outputShape = this->inputs[0]->outputShape;
-    }
+    void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-        auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-        eTVec(output).device(*device) = eTVec(inputs[0]).exp();
-    }
+    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
 
     void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
                      const Tensor<T> *output,
                      const Tensor<T> *outputGradient,
                      size_t index,
-                     Tensor<T> *iGradient) override {
-        DEEP8_ARGUMENT_CHECK(0 == index, "the index of Exp backwardCPU is error");
-
-        auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
-
-        eTVec(iGradient).device(*device) += (eTVec(output) * eTVec(outputGradient));
-    }
-
+                     Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
