+++ deep8/nodes/L1Norm.h
@@ -1,20 +1,11 @@
 #ifndef DEEP8_L1NORM_H
 #define DEEP8_L1NORM_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
-template <typename T>
-struct L1NormBackwardExpr {
-    inline T operator()(T outputGrad, T input) const {
-        if (input > T(0)) {
-            return outputGrad;
-        } else if (input < T(0)) {
-            return -outputGrad;
-        } else {
-            return 0;
-        }
-    }
-};
+
 
 #ifdef HAVE_CUDA
 
@@ -108,47 +99,11 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the L1Norm Function needs only 1 input");
-
-        this->outputShape = Shape({this->inputs[0]->outputShape.batch(), 1});
-    }
+    void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-        auto eigenDevice = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-        auto input = inputs[0];
-        auto batch = input->batch();
-        auto size  = input->size() / batch;
-
-        Eigen::array<size_t, 2> reshapeDims = {batch, size};
-        Eigen::array<size_t, 1> sumDims = {1};
-
-        eTVec(output).device(*eigenDevice) = eTVec(input).abs().reshape(reshapeDims).sum(sumDims);
-    }
-
-    void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-					const Tensor<T> *output,
-					const Tensor<T> *outputGradient,
-					size_t index,
-					Tensor<T> *iGradient) override {
-		DEEP8_ARGUMENT_CHECK(0 == index, "the index of L1Norm backwardCPU is error");
-
-        auto eigenDevice = static_cast<CPUDevice*>(iGradient->device())->eigenDevice;
-
-        auto batch = iGradient->batch();
-        auto size  = iGradient->size() / batch;
-
-        Eigen::array<size_t, 2> iGradientDims = {batch, size};
-        Eigen::array<size_t, 2> outputGradientDims = {batch, 1};
-        Eigen::array<size_t, 2> broadDims = {1, size};
-
-        eTVec(iGradient).reshape(iGradientDims).device(*eigenDevice) +=
-                eTVec(outputGradient).reshape(outputGradientDims).broadcast(broadDims).binaryExpr(eTVec(inputs[0]).reshape(iGradientDims), L1NormBackwardExpr<T>());
-    }
+    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
+    void backwardCPU(const std::vector<const Tensor<T>*> &inputs, const Tensor<T> *output, const Tensor<T> *outputGradient, size_t index, Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
